feat(quartzite-event-types): extract event vocabulary into new no_std crate
Create `quartzite-event-types` holding `Event<T>`, `EventType<T>`,
`EventFilter<T>`, `KeyEventKind`, `MouseEventKind`, and `TimerEvent`.
`quartzite-events` re-exports all six so existing use-paths compile
unchanged. `quartzite-runtime` depends on the new crate instead of
`quartzite-events` and switches `Timer.tick` from `Signal<(usize,)>`
to `Signal<(TimerEvent,)>`; `TimerEvent` gains a `fire_count: usize`
field so no information is lost.
4 new tests in `quartzite-event-types`; all 679 tests green.
