+++ quartzite-event-types/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "quartzite-event-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Core event vocabulary (Event, EventType, EventFilter, TimerEvent) for quartzite — no_std + alloc"
+
+# No [features] — always no_std + alloc
+
+[dependencies]
+quartzite-core = { path = "../quartzite-core", default-features = false }
+
+[dev-dependencies]
+rstest            = "0.26"
+pretty_assertions = "1"
+
+[package.metadata.docs.rs]
+rustdoc-args = ["--cfg", "docsrs"]
+rustc-args = ["--cfg", "docsrs"]
