+++ quartzite-runtime/Cargo.toml
@@ -9,12 +9,13 @@ repository.workspace = true
 description = "Application runtime, event loop, and object tree for quartzite"
 
 [dependencies]
-parking_lot = "0.12"
-quartzite-core = { path = "../quartzite-core" }
-quartzite-macros = { path = "../quartzite-macros" }
-slotmap = "1"
-thiserror = "2"
-tracing = { version = "0.1", features = ["log"] }
+parking_lot           = "0.12"
+quartzite-core        = { path = "../quartzite-core" }
+quartzite-event-types = { path = "../quartzite-event-types" }
+quartzite-macros      = { path = "../quartzite-macros" }
+slotmap               = "1"
+thiserror             = "2"
+tracing               = { version = "0.1", features = ["log"] }
 
 [dev-dependencies]
 env_logger = "0.11"
